Remove unused exports from Browser System.Native build
