+++ src/libraries/System.IO.FileSystem/src/System.IO.FileSystem.csproj
@@ -6,6 +6,9 @@
     <TargetFrameworks>$(NetCoreAppCurrent)-Windows_NT;$(NetCoreAppCurrent)-Unix;$(NetCoreAppCurrent)-Browser</TargetFrameworks>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+  <PropertyGroup Condition="'$(TargetsBrowser)' == 'true'">
+    <DefineConstants>$(DefineConstants);TARGET_BROWSER</DefineConstants>
+  </PropertyGroup>
   <ItemGroup>
     <!-- System.IO.Enumeration.FileSystemEntry is not correctly parsed by CsharpAnalyser so we are skipping this type from autogenerated typeforwards file.  -->
     <GenFacadesIgnoreMissingType Include="System.IO.Enumeration.FileSystemEntry" />
@@ -178,8 +181,6 @@
              Link="Interop\Unix\Interop.Errors.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\Interop.IOErrors.cs"
              Link="Interop\Unix\Interop.IOErrors.cs" />
-    <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.GetEUid.cs"
-             Link="Common\Interop\Unix\Interop.GetEUid.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.MountPoints.cs"
              Link="Common\Interop\Unix\Interop.MountPoints.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.Permissions.cs"
@@ -202,8 +203,6 @@
              Link="Common\Interop\Unix\Interop.ChMod.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.CopyFile.cs"
              Link="Common\Interop\Unix\Interop.CopyFile.cs" />
-    <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.GetEGid.cs"
-             Link="Common\Interop\Unix\Interop.GetEGid.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.LChflags.cs"
              Link="Common\Interop\Unix\Interop.LChflags.cs" />
     <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.Link.cs"
@@ -226,6 +225,12 @@
     <Compile Include="System\IO\FileSystem.Exists.Unix.cs" />
     <Compile Include="System\IO\FileSystem.Unix.cs" />
     <Compile Include="System\IO\FileSystemInfo.Unix.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetsUnix)' == 'true'">
+    <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.GetEUid.cs"
+             Link="Common\Interop\Unix\Interop.GetEUid.cs" />
+    <Compile Include="$(CommonPath)Interop\Unix\System.Native\Interop.GetEGid.cs"
+             Link="Common\Interop\Unix\Interop.GetEGid.cs" />
   </ItemGroup>
   <!-- Unix references -->
   <ItemGroup Condition="'$(TargetsUnix)' == 'true' or '$(TargetsBrowser)' == 'true'">
